New DFA-based matching algorithm for the C scanner.
This has the major advantage of being completely precalculated, which
means the library has no state, makes no memory allocations, and does
not need to be initialized.  The DFA fits in approximately 43K of
read-only data space, which is a little less than half of the
*runtime* memory requirement for the old code, ignoring the size of
that code's read-only data (which IIRC was about 20K).  It should also
be considerably faster, but I haven't benchmarked it yet.

Note that I am checking in a copy of effective_tld_names.dat in this
commit, so that the build process (which will be completely automated
shortly) need not reach out to the network.

Note also that tld-canon.h is now generated via a Python program rather
than the old PHP program.  The old code is still present, but will be
removed shortly.

 * compile_tlds.py: New program that generates DFAs to drive the new
   matching algorithm.
 * effective_tld_names.dat: Copy from Mozilla
   (upstream hg revision e685ff7fa126).

 * C/regdom.c: Rewrite with new matching algorithm.
 * C/tld-canon.h: Regenerate using compile_tlds.py.

 * C/regdom.h: Public API now consists only of getRegisteredDomain and
   getRegisteredDomainDrop.
 * C/test-regdom.c: Update to match.
 * C/Makefile: -Wno-unused-parameter and -Wno-overlength-strings are
   no longer required.
 * C/punycode.pl: Removed.